Update Sudoku::Container module
Turns out, ruby mixins don't like instance variables. Therefore, we'll
interact with the member_squares array through get_members and
set_member methods. We'll also use a calc_members method to put the
functionality of deciding which squares should belong to a certain
container with said container, not with the board object.