Create NullSquare class
NullSquare follows the Null Object Pattern and provides a square object
that returns null/raises exceptions for methods the square class is
expected to support. This will allow collections of squares to be
iterated over if some of the enumerable item's aren't set. (i.e. a
collection of squares that skip the zero index so that IDs and indexes
line up.