Change Container#set_method name
We are no long putting a member square at it's id's index because this
caused problems when trying to iterate over the collection. Therefore,
we are now pushing new members onto the array. Changing the method name
to add_member from set_member seems to better convey what's actually
going on.