Update Board.squares and baard.get_members
Adds a NullSquare object to the zero index of board.squares. This allows
board.squares to be iterated over and keeps the square.id in line with
the board.squares index.

Updates board.get_members, changing the method name to
set_container_members and utilizing the container interface/mixin by
calling the container.calc_members method and using the
container.add_member method to add member squares to each container.