Update Board.get_members calls in Column
The method Board#get members was updated to a more descriptive
set_container_members. This commit updates Column's calls to this
method.

The method of storing member_squares in Column has been altered from
storing them in the index that matches their id to a stack method. This
makes it much easier to iterate over Column.member_squares.