Update Sudoku::Container#update_possible_values
We already handle sending the Sudoku::Board#illegal_move message in
Sudoku::Square#update. No need to do it again. Also, I like the idea of
having a Move object to model the state of each move. This would mean we
wouldn't need to the square to the illegal_move method because the board
will know which move is current. Therefore, we removed the square
argument from the illegal_move call in Sudoku::Square#illegal_move.